requirements.txt
openai>=1.30.0
streamlit>=1.35.0
python-dotenv>=1.0.1
pandas>=2.2.2
requests>=2.32.3
